rename LIController and LIDarkNews to SIREN_*
