+++ resources/Examples/Example2/DipolePortal_CCM.py
@@ -2,7 +2,7 @@ import os
 import numpy as np
 
 import siren
-from siren.LIController import LIController
+from siren.SIREN_Controller import SIREN_Controller
 
 # Define a DarkNews model
 model_kwargs = {
@@ -24,7 +24,7 @@ events_to_inject = 100000
 experiment = "CCM"
 
 # Define the controller
-controller = LIController(events_to_inject, experiment)
+controller = SIREN_Controller(events_to_inject, experiment)
 
 # Particle to inject
 primary_type = siren.dataclasses.Particle.ParticleType.NuMu
